Make single script for bot. Multiplayer still needs testing.
