+++ .vscode/settings.json
@@ -4,7 +4,9 @@
     "**/__pycache__": true,
     "**/__pycache__.*": true,
     "**/.pytest_cache": true,
-    "**/.pytest_cache.*": true
+    "**/.pytest_cache.*": true,
+    "**/build": true,
+    "**/build.*": true
   },
   "python.analysis.typeCheckingMode": "basic",
   "python.analysis.autoImportCompletions": true
