+++ README.md
@@ -7,7 +7,7 @@
 ## How to run
 
 - `pip install .`
-- `TIC_TAC_TOE_TOKEN_TG=token app`
+- `TIC_TAC_TOE_TOKEN_TG=token app` (entry point) or `python -m tic_tac_toe.bot`
 
 ## TODO
 
@@ -17,7 +17,7 @@
 - [x] Add bot logic
 - [x] Understand `python-telegram-bot` architecture (`update`, `context`)
 - [x] Add multiplayer
-- [ ] Refactor project (single main script, readable code)
+- [ ] Refactor project (single main script, structured code)
 - [x] Make better exception hierarchy
 - [ ] Add some description, annotations, docstrings
 - [ ] Make more interesting text
